# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/automation/run": {
      "post": {
        "operationId": "AutomationController_run",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/automation/image": {
      "post": {
        "operationId": "AutomationController_image",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/automation/images": {
      "get": {
        "operationId": "AutomationController_listImages",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/automation/images/{id}": {
      "delete": {
        "operationId": "AutomationController_deleteImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "ProdPilot API",
    "description": "AIpreneur Content Engine — kontrak API v1 untuk web Next.js",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AutomateDto": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string"
          },
          "brief": {
            "type": "string",
            "description": "Brief / arahan konten dari user (jadi dasar spec JSON).",
            "minLength": 1
          },
          "format": {
            "type": "string",
            "description": "Bentuk output yang mau dibuat.",
            "enum": [
              "reels",
              "carousel",
              "post"
            ]
          },
          "platform": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "anthropic",
              "openai"
            ]
          },
          "apiKey": {
            "type": "string"
          },
          "model": {
            "type": "string"
          }
        },
        "required": [
          "brief",
          "format"
        ]
      },
      "GenerateImageDto": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string"
          },
          "brief": {
            "type": "string",
            "description": "Brief / arahan visual dari user.",
            "minLength": 1
          },
          "count": {
            "type": "number",
            "description": "Jumlah gambar/konsep yang dibuat (1–4, default 3).",
            "minimum": 1,
            "maximum": 4
          },
          "aspect": {
            "type": "string",
            "description": "Rasio gambar (default square).",
            "enum": [
              "square",
              "portrait",
              "landscape"
            ]
          },
          "style": {
            "type": "string",
            "description": "Gaya visual opsional (mis. \"flat vector\", \"fotografi produk\", \"3D render\")."
          },
          "apiKey": {
            "type": "string",
            "description": "BYO OpenAI key dari UI (image hanya via OpenAI); tidak disimpan."
          },
          "model": {
            "type": "string",
            "description": "Override model image (mis. dall-e-3 / gpt-image-1)."
          }
        },
        "required": [
          "brief"
        ]
      }
    }
  }
}